+++ src/services/src/Demoulas.ProfitSharing.Services/ProfitShareUpdate/MemberFinancials.cs
@@ -18,15 +18,15 @@ public class MemberFinancials
     public decimal Military { get; set; }
     public decimal Xfer { get; set; }
     public decimal Pxfer { get; set; }
-    public long EmployeeTypeId { get; set; }
-    public long ContributionPoints { get; set; }
-    public long EarningPoints { get; set; }
+    public byte EmployeeTypeId { get; set; }
+    public int ContributionPoints { get; set; }
+    public int EarningPoints { get; set; }
     public decimal Contributions { get; set; }
     public decimal IncomingForfeitures { get; set; }
     public decimal Earnings { get; set; }
     public decimal SecondaryEarnings { get; set; }
     public decimal MaxOver { get; set; }
-    public long MaxPoints { get; set; }
+    public int MaxPoints { get; set; }
     public long Psn { get; set; }
     public decimal Caf { get; set; }
     public decimal EndingBalance { get; set; }
